async/await, io renaming and fix variable.value
variable.value was missing a type setting on the value out.

Some of the inputs and outputs have been renamed
to address the fact that they can obtain and return multiple values.

The variable node itself has an output to return the values with the correct type.
Note that variable.value has an object type set so this needs casting.
The values out on the variable node does not need casting.