closes #80, closes #12
This introduces improved error handling which is both more user friendly 
to an end-user of the UI, as it is to a developer of nodepacks.
Throwing an error within a nodepack for instance now shows up properly 
in the UI, and those errors can be caught by a 
xible.flow.instance.on-error node.