centralize flow start, deploy and delete commands
As a result, when starting a flow from the flows overview, users will be 
prompted for xible.flow.instance.param values as well.
Previously only the deploy and start buttons in the editor handled this.

Also, flows are now properly removed from the editor flow list, when a 
flow is deleted remotely.